requirements.txt
rdflib
jsonpath_ng
requests
pyyaml
jsonschema
pandas
pyshacl
pyld
